+++ test/mocks/ssl/mocks.h
@@ -39,6 +39,7 @@ public:
   ~MockConnectionInfo() override;
 
   MOCK_METHOD(bool, peerCertificatePresented, (), (const));
+  MOCK_METHOD(bool, peerCertificateValidated, (), (const));
   MOCK_METHOD(absl::Span<const std::string>, uriSanLocalCertificate, (), (const));
   MOCK_METHOD(const std::string&, sha256PeerCertificateDigest, (), (const));
   MOCK_METHOD(const std::string&, serialNumberPeerCertificate, (), (const));